+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 ### [@project-version@](https://github.com/wow-addon-dev/Horatum/tree/@project-version@) (@build-date@)
 [Full Changelog](@full-changelog@) - [Previous Releases](https://github.com/wow-addon-dev/Horatum/releases)
 
-- Added: The boss combat encouters in story mode raids are now being tracked
+- Added: Instance type is displayed in the Combat Time Tracker